refactor: scss 변경
