+++ tovalley-react/package.json
@@ -22,6 +22,7 @@
     "chart.js": "^4.3.3",
     "luxon": "^3.4.4",
     "moment": "^2.29.4",
+    "node-sass": "^7.0.3",
     "react": "^18.2.0",
     "react-chartjs-2": "^5.2.0",
     "react-cookie": "^6.0.1",
