+++ tovalley-react/src/styles/home/home.module.scss
@@ -1,0 +1,29 @@
+.mainPage {
+  min-width: 360px;
+  margin-top: 120px;
+}
+
+.body {
+  padding: 0 7%;
+}
+
+.top {
+  display: flex;
+  margin-bottom: 3em;
+}
+
+.lostContents {
+  display: flex;
+}
+
+.recentPost {
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 5em;
+}
+
+@media screen and (max-width: 1000px) {
+  .top {
+    flex-direction: column;
+  }
+}
